fix: move pip chat to conference menu
